fix path detection (again)
